@@ -1,5 +1,5 @@
-﻿using Microsoft.AspNetCore;
-using Microsoft.AspNetCore.Hosting;
+﻿using Microsoft.AspNetCore.Hosting;
+using Microsoft.Extensions.Hosting;
 using System.IO;
 
 namespace Trainer
@@ -8,18 +8,15 @@ namespace Trainer
     {
         public static void Main(string[] args)
         {
-            CreateWebHostBuilder(args).Build().Run();
+            CreateHostBuilder(args).Build().Run();
         }
 
-        public static IWebHostBuilder CreateWebHostBuilder(string[] args) =>
-         WebHost.CreateDefaultBuilder(args)
-           .UseContentRoot(Directory.GetCurrentDirectory())
-           .UseStartup<Startup>()
-           .PreferHostingUrls(true);
-
-        //.ConfigureAppConfiguration((hostingContext, config) =>
-        //{
-        //    config.AddJsonFile("emailTemplates.json", optional: false, reloadOnChange: false);
-        //})
+        public static IHostBuilder CreateHostBuilder(string[] args) =>
+         Host.CreateDefaultBuilder(args).ConfigureWebHostDefaults(webBuilder =>
+         {
+             webBuilder.UseContentRoot(Directory.GetCurrentDirectory());
+             webBuilder.UseStartup<Startup>();
+             webBuilder.PreferHostingUrls(true);
+         });
     }
 }
